chore: make every algorithm single file(almost, except for kmeans++)
